Add spec for rm

+++ lib/switch_db/command/rm.rb
@@ -1,13 +1,14 @@
+require 'fileutils'
+
 module SwitchDb
   module Command
-    class Rm
-      attr_reader :reference
+    class Rm < Base
+      def run(name:, **)
+        reference = @reference_set.references[name.to_s]
+        return unless reference
 
-      def initialize(reference)
-        @reference = reference
-      end
-
-      def rm!
+        @reference_set.remove_reference(reference)
+        puts "Deleted reference a (#{reference.name})"
         FileUtils.rm_rf(reference.full_path)
       end
     end
